requirements.txt
Flask==3.0.1
gunicorn==23.0.0

requests==2.32.0
structlog==24.1.0

google-auth==2.3.2

discord.py
python-dotenv
aiosqlite
psycopg2 
langchain
langchain_community
langchain_openai
langchain-ollama
flake8
black
isort
pygithub
langgraph
PyJWT
cryptography
fastapi
uvicorn
ruff
